requirements.txt
google-generativeai
python-dotenv
tkcalendar
tkinterweb markdown2